+++ src/functions/CrossProduct.java
@@ -2,12 +2,12 @@ package functions;
 
 public class CrossProduct {
 	
-	float[] CrossPrdct(float[] a, float[] b) {		//Produto vetorial
+	double[] CrossPrdct(double[] a, double[] b) {		//Produto vetorial
 		if(a.length != b.length || a.length != 3) {
 			System.out.println("Vetores mal definidos");
 			return null;
 		} else {
-			float[] result = new float[3];
+			double[] result = new double[3];
 			
 			result[0] = a[1] * b[2] + (-1)*(a[2] * b[1]);
 			result[1] = a[2] * b[0] + (-1)*(a[0] * b[2]);
