+++ src/functions/MatricesMult.java
@@ -2,14 +2,14 @@ package functions;
 
 public class MatricesMult {
 	
-	public float[][] MatricesMltply(float[][] a, float[][] b){
+	public double[][] MatricesMltply(double[][] a, double[][] b){
 		if(a[0].length != b.length) {
 			System.out.println("Matrizes não podem ser multiplicadas");
 			return null;
 		}
 		else {
-			float[][] result = new float[a.length][b[0].length];
-			float sum = 0;
+			double[][] result = new double[a.length][b[0].length];
+			double sum = 0;
 			
 			for(int i=0; i < a.length; i++) {			//i linha do primeiro
 				for(int l=0; l < b[0].length; l++) {	//l coluna do segundo
